Add files via upload
A landing page is a single web page that serves as the entry point for a website or a particular section of a website. It is designed to target a specific audience and encourage visitors to take a particular action, such as signing up for a service, downloading a resource, or making a purchase. Here's what you might find in a landing page project on GitHub:

1. HTML Structure: The landing page project typically includes one or more HTML files that define the structure of the page. These files contain the necessary markup to display text, images, forms, and other elements.
2. CSS Styling: The project includes one or more CSS files that define the styles for the landing page. These files control the layout, colors, fonts, and other visual aspects of the page to create an attractive and cohesive design.